api/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
sqlalchemy==2.0.35
psycopg[binary,pool]==3.2.1
beautifulsoup4==4.12.3
trafilatura==1.12.2
python-dotenv==1.0.0
pandas==2.2.2
